Implement a quit command.
This replaces the "exit if the q key is pressed" temporary logic with a mode system. The modes do different a bit from vim: this editor has "command" mode (where pressing individual keys or short sequences triggers behaviors) and "prompt" mode (where you type a command line). vim calls these both "command".

For now, only the q and q! commands are implemented (and q doesn't check if the current file is saved, because that concept does not exist yet).